requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
install==1.3.5
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
